change unit stock and make payment method not requirement in first post order

+++ app/src/main/java/com/example/wp/data/api/datasource/MaterialRepositoryImpl.kt
@@ -72,7 +72,7 @@ class MaterialRepositoryImpl(private val service: MaterialService) : MaterialRep
 
     override suspend fun updateMaterial(
         materialId: Int,
-        stock: Int,
+        stock: Double,
         type: String,
         reason: String
     ): Load<Boolean> {
