+++ app/src/main/java/com/example/wp/data/api/model/request/RequestMaterialApi.kt
@@ -1,12 +1,12 @@
 package com.example.wp.data.api.model.request
 
 data class RequestMaterialApi(
-    val stock: Int,
+    val stock: Double,
     val material: String
 )
 
 data class RequestQuantityMaterialApi(
-    val stock: Int,
+    val stock: Double,
     val type: String,
     val reason:String
 )
